requirements.txt
accelerate==1.2.1
annotated-types==0.7.0
av==13.1.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
diffusers==0.31.0
docker-pycreds==0.4.0
einops==0.8.0
filelock==3.16.1
fsspec==2024.12.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.27.0
idna==3.10
importlib_metadata==8.5.0
jaxtyping==0.2.36
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.1
opencv-python==4.10.0.84
packaging==24.2
peft==0.14.0
pillow==11.0.0
platformdirs==4.3.6
protobuf==5.29.2
psutil==6.1.1
pydantic==2.10.4
pydantic_core==2.27.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
sentry-sdk==2.19.2
setproctitle==1.3.4
six==1.17.0
smmap==5.0.1
sympy==1.13.1
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.47.1
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.3.0
wandb==0.19.1
zipp==3.21.0
huggingface_hub